+++ Packages/com.yontalane.dialog/Editor/ScriptableDialogAgentEditor.cs
@@ -17,6 +17,7 @@ namespace YontalaneEditor.Dialog
         private SerializedProperty m_staticText;
         private SerializedProperty m_keywords;
         private SerializedProperty m_inlineImageReplacementInfo;
+        private SerializedProperty m_lineBuilder;
         private GUIContent m_dataLabel;
         private GUIContent m_textDataStartLabel;
         private GUIContent m_textDataLabel;
@@ -34,6 +35,7 @@ namespace YontalaneEditor.Dialog
             m_staticText = serializedObject.FindProperty("m_staticText");
             m_keywords = serializedObject.FindProperty("m_keywords");
             m_inlineImageReplacementInfo = serializedObject.FindProperty("m_inlineImageReplacementInfo");
+            m_lineBuilder = serializedObject.FindProperty("m_lineBuilder");
 
             m_dataLabel = new GUIContent(string.Empty, m_data.tooltip);
             m_textDataStartLabel = new GUIContent("Starting Node", m_textDataStart.tooltip);
@@ -69,6 +71,7 @@ namespace YontalaneEditor.Dialog
 
             EditorGUILayout.PropertyField(m_keywords);
             EditorGUILayout.PropertyField(m_inlineImageReplacementInfo);
+            EditorGUILayout.PropertyField(m_lineBuilder);
 
             serializedObject.ApplyModifiedProperties();
         }